update for v4
v4 doesn't have any new settings that most people will want to
configure, so I haven't made them part of the env config for Docker.

This switches to docker compose v2 (see
https://docs.docker.com/compose/migrate/); the only thing to note is
that `docker-compose` is now `docker compose`.

I've changed the containers that use the LTI provider image to refer to
iy by name, and the build step tags the container as
`numbas-lti-provider`.